Tests/LibWeb/Text/expected/geolocation-emulated.txt

latitude: 37.7647658
longitude: -122.4345892
accuracy type: number
timestamp type: number
altitude: 0
